Move code to PlutoRunner
